correct typing

+++ code/FINDER_ND_cost/PrepareBatchGraph.pyx
@@ -136,8 +136,8 @@ cdef class py_PrepareBatchGraph:
         colNum= matrix.colNum
         indices = np.mat([rowIndex, colIndex]).transpose()
 
-        index = torch.tensor(np.transpose(np.array(indices)))
-        value = torch.tensor(np.array(data))
+        index = torch.tensor(np.transpose(np.array(indices)), dtype=torch.double)
+        value = torch.tensor(np.array(data), dtype=torch.double)
         index, value = torch_sparse.coalesce(index, value, m=rowNum, n=colNum)
         return_dict = {"index": index, "value": value, "m":rowNum, "n":colNum}
